Rename vim to nvim
I've been using neovim for a while now and it's just better.
So in this change I'm renaming my "vim" directory to "nvim",
and renaming "vimrc" to "init.vim". The next step would be
to rename "init.vim" to "init.lua", and use Lua, which
I've started to do [here](https://git.io/JPHGH), but
I'm still not convinced that this is going to improve
anything as far as speed or readability is concerned.

I might do it one day. In the meantime, at least my
vim configuration is stored in ~/.config/nvim which
is nicer than in ~/.vim.